+++ app/services/carto/user_metadata_export_service.rb
@@ -1,15 +1,25 @@
 require 'json'
 require 'carto/export/layer_exporter'
 require 'carto/export/data_import_exporter'
+require_dependency 'carto/export/connector_configuration_exporter'
+
+# Not migrated
+# client_applications & friends -> deprecated?
+# likes -> difficult to do between clouds
+# snapshots -> difficult to do between clouds, not in use yet
+# tags -> regenerated from tables
+# visualization_export -> only purpose would be logging
 
 # Version History
 # 1.0.0: export user metadata
 # 1.0.1: export search tweets
+# 1.0.2: export user notifications
 # 1.0.3: export rate limits
 # 1.0.4: company and phone in users table
+# 1.0.5: synchronization_oauths and connector configurations
 module Carto
   module UserMetadataExportServiceConfiguration
-    CURRENT_VERSION = '1.0.4'.freeze
+    CURRENT_VERSION = '1.0.5'.freeze
     EXPORTED_USER_ATTRIBUTES = [
       :email, :crypted_password, :salt, :database_name, :username, :admin, :enabled, :invite_token, :invite_token_date,
       :map_enabled, :quota_in_bytes, :table_quota, :account_type, :private_tables_enabled, :period_end_date,
@@ -40,6 +50,7 @@ module Carto
     include UserMetadataExportServiceConfiguration
     include LayerImporter
     include DataImportImporter
+    include ConnectorConfigurationImporter
 
     def build_user_from_json_export(exported_json_string)
       build_user_from_hash_export(parse_json(exported_json_string))
@@ -93,7 +104,13 @@ module Carto
       api_keys = exported_user[:api_keys] || []
       user.api_keys += api_keys.map { |api_key| Carto::ApiKey.new_from_hash(api_key) }
 
-      user.static_notifications = Carto::UserNotification.create(notifications: exported_user[:notifications])
+      if exported_user[:notifications]
+        user.static_notifications = Carto::UserNotification.create(notifications: exported_user[:notifications])
+      end
+
+      user.synchronization_oauths = build_synchronization_oauths_from_hash(exported_user[:synchronization_oauths])
+
+      user.connector_configurations = build_connector_configurations_from_hash(exported_user[:connector_configurations])
 
       # Must be the last one to avoid attribute assignments to try to run SQL
       user.id = exported_user[:id]
@@ -137,12 +154,28 @@ module Carto
 
       rate_limit
     end
+
+    def build_synchronization_oauths_from_hash(exported_array)
+      return [] unless exported_array.present?
+
+      exported_array.map { |so| build_synchronization_oauth_from_hash(so) }
+    end
+
+    def build_synchronization_oauth_from_hash(exported_hash)
+      SynchronizationOauth.new(
+        service: exported_hash[:service],
+        token: exported_hash[:token],
+        created_at: exported_hash[:created_at],
+        updated_at: exported_hash[:updated_at]
+      )
+    end
   end
 
   module UserMetadataExportServiceExporter
     include UserMetadataExportServiceConfiguration
     include LayerExporter
     include DataImportExporter
+    include ConnectorConfigurationExporter
 
     def export_user_json_string(user)
       export_user_json_hash(user).to_json
@@ -173,6 +206,12 @@ module Carto
       user_hash[:rate_limit] = export_rate_limit(user.rate_limit)
 
       user_hash[:notifications] = user.static_notifications.notifications
+
+      user_hash[:synchronization_oauths] = user.synchronization_oauths.map { |so| export_synchronization_oauth(so) }
+
+      user_hash[:connector_configurations] = user.connector_configurations.map do |cc|
+        export_connector_configuration(cc)
+      end
 
       user_hash
     end
@@ -216,6 +255,15 @@ module Carto
       {
         id: rate_limit.id,
         limits: rate_limit.api_attributes
+      }
+    end
+
+    def export_synchronization_oauth(sync_oauth)
+      {
+        service: sync_oauth.service,
+        token: sync_oauth.token,
+        created_at: sync_oauth.created_at,
+        updated_at: sync_oauth.updated_at
       }
     end
   end